backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
supabase==2.18.1
pydantic-settings==2.10.1
python-dotenv==1.1.1
bcrypt==4.1.1
PyJWT==2.11.0